feat: useIsInProgress 개발 및 적용
